+++ src/interfaces/multisigCalls.ts
@@ -1,0 +1,11 @@
+export interface multisig_calls {
+    multisig_address: string,
+    call_hash: string,
+    call_data: string,
+    status: string,
+    approvals: string[],
+    depositor: string,
+    deposit: string,
+    when: any,
+    chain: string
+}
